requirements.txt
flask
requests
openai==0.28.1
python-dotenv
beautifulsoup4